requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
torch>=1.9.0
transformers>=4.12.0
accelerate>=0.26.0
huggingface_hub>=0.0.8
pydantic>=1.8.0
python-dotenv
